+++ examples/basic/index.jsx
@@ -5,14 +5,14 @@ import thunk from 'redux-thunk';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-const inputConfig = {
+const inputsConfig = {
     email: {
         defaultValue: 'test@example.com',
         validator: (value) => typeof value === 'string' && value.indexOf('@') >= 0
     }
 };
 const reducer = combineReducers({
-    inputs: createInputsReducer(inputConfig)
+    inputs: createInputsReducer(inputsConfig)
 });
 const store = createStore(reducer, applyMiddleware(thunk));
 
@@ -25,14 +25,25 @@ let EmailInput = ({id, value, error, onChange}) => (
 EmailInput = ReduxInputsWrapper(EmailInput);
 
 function Form(props) {
-    const { inputs, inputProps } = props;
+    const { inputs, reduxInputs } = props;
+    const { inputProps, values, valid, pristine, validating } = reduxInputs;
     return (
         <form>
             <EmailInput {...inputProps.email}/>
+
+            <h3>reduxInputs prop</h3>
+            <ul>
+                <li>values: <pre>{JSON.stringify(values, null, 2)}</pre></li>
+                <li>valid: {valid.toString()}</li>
+                <li>pristine: {pristine.toString()}</li>
+                <li>validating: {validating.toString()}</li>
+            </ul>
+
             <h3>Input state</h3>
             <pre>{JSON.stringify(inputs, null, 2)}</pre>
+
         </form>
     );
 }
-const FormContainer = connectWithInputs(inputConfig)(s => s)(Form);
+const FormContainer = connectWithInputs(inputsConfig)(s => s)(Form);
 ReactDOM.render(<Provider store={store}><FormContainer /></Provider>, document.getElementById('container'));